Separate core and ui for GetterSetterCorrectionSubProcessor and VarargsWarningsSubProcessor
