+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/correction/VarargsWarningsSubProcessor.java
@@ -16,7 +16,6 @@
 package org.eclipse.jdt.internal.ui.text.correction;
 
 import java.util.Collection;
-import java.util.List;
 
 import org.eclipse.swt.graphics.Image;
 
@@ -24,30 +23,17 @@ import org.eclipse.ui.ISharedImages;
 import org.eclipse.ui.PlatformUI;
 
 import org.eclipse.jdt.core.ICompilationUnit;
-import org.eclipse.jdt.core.JavaModelException;
-import org.eclipse.jdt.core.dom.ASTNode;
-import org.eclipse.jdt.core.dom.ClassInstanceCreation;
-import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.IMethodBinding;
-import org.eclipse.jdt.core.dom.ITypeBinding;
-import org.eclipse.jdt.core.dom.MarkerAnnotation;
 import org.eclipse.jdt.core.dom.MethodDeclaration;
-import org.eclipse.jdt.core.dom.MethodInvocation;
-import org.eclipse.jdt.core.dom.Modifier;
 import org.eclipse.jdt.core.dom.rewrite.ASTRewrite;
 
-import org.eclipse.jdt.internal.core.manipulation.dom.ASTResolving;
-import org.eclipse.jdt.internal.corext.util.JavaModelUtil;
-import org.eclipse.jdt.internal.corext.util.Messages;
-
-import org.eclipse.jdt.ui.text.java.IInvocationContext;
 import org.eclipse.jdt.ui.text.java.correction.ASTRewriteCorrectionProposal;
 import org.eclipse.jdt.ui.text.java.correction.ICommandAccess;
 
 import org.eclipse.jdt.internal.ui.JavaPluginImages;
 import org.eclipse.jdt.internal.ui.text.correction.proposals.LinkedCorrectionProposal;
 
-public class VarargsWarningsSubProcessor {
+public class VarargsWarningsSubProcessor extends VarargsWarningsBaseSubProcessor<ICommandAccess> {
 
 	private static class AddSafeVarargsProposal extends LinkedCorrectionProposal {
 		public AddSafeVarargsProposal(String label, ICompilationUnit cu, MethodDeclaration methodDeclaration, IMethodBinding methodBinding, int relevance) {
@@ -55,89 +41,32 @@ public class VarargsWarningsSubProcessor {
 		}
 	}
 
-	public static void addAddSafeVarargsProposals(IInvocationContext context, IProblemLocationCore problem, Collection<ICommandAccess> proposals) {
-		ASTNode coveringNode= problem.getCoveringNode(context.getASTRoot());
-
-		MethodDeclaration methodDeclaration= ASTResolving.findParentMethodDeclaration(coveringNode);
-		if (methodDeclaration == null)
-			return;
-
-		IMethodBinding methodBinding= methodDeclaration.resolveBinding();
-		if (methodBinding == null)
-			return;
-
-		int modifiers= methodBinding.getModifiers();
-		if (!Modifier.isStatic(modifiers) && !Modifier.isFinal(modifiers) && !Modifier.isPrivate(modifiers) && !methodBinding.isConstructor())
-			return;
-
-		String label= CorrectionMessages.VarargsWarningsSubProcessor_add_safevarargs_label;
-		AddSafeVarargsProposal proposal= new AddSafeVarargsProposal(label, context.getCompilationUnit(), methodDeclaration, null, IProposalRelevance.ADD_SAFEVARARGS);
-		proposals.add(proposal);
+	public static void addAddSafeVarargsProposals(IInvocationContextCore context, IProblemLocationCore problem, Collection<ICommandAccess> proposals) {
+		new VarargsWarningsSubProcessor().createAddSafeVarargsProposals(context, problem, proposals);
 	}
 
-	public static void addAddSafeVarargsToDeclarationProposals(IInvocationContext context, IProblemLocationCore problem, Collection<ICommandAccess> proposals) {
-		if (!JavaModelUtil.is1d7OrHigher(context.getCompilationUnit().getJavaProject()))
-			return;
-
-		ASTNode coveringNode= problem.getCoveringNode(context.getASTRoot());
-		IMethodBinding methodBinding;
-		if (coveringNode instanceof MethodInvocation) {
-			methodBinding= ((MethodInvocation) coveringNode).resolveMethodBinding();
-		} else if (coveringNode instanceof ClassInstanceCreation) {
-			methodBinding= ((ClassInstanceCreation) coveringNode).resolveConstructorBinding();
-		} else {
-			return;
-		}
-		if (methodBinding == null)
-			return;
-
-		String label= Messages.format(CorrectionMessages.VarargsWarningsSubProcessor_add_safevarargs_to_method_label, methodBinding.getName());
-
-		ITypeBinding declaringType= methodBinding.getDeclaringClass();
-		CompilationUnit astRoot= (CompilationUnit) coveringNode.getRoot();
-		if (declaringType != null && declaringType.isFromSource()) {
-			try {
-				ICompilationUnit targetCu= ASTResolving.findCompilationUnitForBinding(context.getCompilationUnit(), astRoot, declaringType);
-				if (targetCu != null) {
-					AddSafeVarargsProposal proposal= new AddSafeVarargsProposal(label, targetCu, null, methodBinding.getMethodDeclaration(), IProposalRelevance.ADD_SAFEVARARGS);
-					proposals.add(proposal);
-				}
-			} catch (JavaModelException e) {
-				return;
-			}
-		}
+	public static void addAddSafeVarargsToDeclarationProposals(IInvocationContextCore context, IProblemLocationCore problem, Collection<ICommandAccess> proposals) {
+		new VarargsWarningsSubProcessor().createAddSafeVarargsToDeclarationProposals(context, problem, proposals);
 	}
 
-	public static void addRemoveSafeVarargsProposals(IInvocationContext context, IProblemLocationCore problem, Collection<ICommandAccess> proposals) {
-		ASTNode coveringNode= problem.getCoveringNode(context.getASTRoot());
-		if (!(coveringNode instanceof MethodDeclaration))
-			return;
-
-		MethodDeclaration methodDeclaration= (MethodDeclaration) coveringNode;
-		MarkerAnnotation annotation= null;
-
-		for (ASTNode node : (List<? extends ASTNode>) methodDeclaration.modifiers()) {
-			if (node instanceof MarkerAnnotation) {
-				annotation= (MarkerAnnotation) node;
-				if ("SafeVarargs".equals(annotation.resolveAnnotationBinding().getName())) { //$NON-NLS-1$
-					break;
-				}
-			}
-		}
-
-		if (annotation == null)
-			return;
-
-		ASTRewrite rewrite= ASTRewrite.create(coveringNode.getAST());
-		rewrite.remove(annotation, null);
-
-		String label= CorrectionMessages.VarargsWarningsSubProcessor_remove_safevarargs_label;
-		Image image= PlatformUI.getWorkbench().getSharedImages().getImage(ISharedImages.IMG_TOOL_DELETE);
-		ASTRewriteCorrectionProposal proposal= new ASTRewriteCorrectionProposal(label, context.getCompilationUnit(), rewrite, IProposalRelevance.REMOVE_SAFEVARARGS, image);
-		proposals.add(proposal);
+	public static void addRemoveSafeVarargsProposals(IInvocationContextCore context, IProblemLocationCore problem, Collection<ICommandAccess> proposals) {
+		new VarargsWarningsSubProcessor().createRemoveSafeVarargsProposals(context, problem, proposals);
 	}
 
 	private VarargsWarningsSubProcessor() {
 	}
 
+	@Override
+	protected ICommandAccess createAddSafeVarargsProposal1(String label, ICompilationUnit compilationUnit, MethodDeclaration methodDeclaration, Object object, int addSafevarargs) {
+		return new AddSafeVarargsProposal(label, compilationUnit, methodDeclaration, null, IProposalRelevance.ADD_SAFEVARARGS);
+	}
+	@Override
+	protected ICommandAccess createAddSafeVarargsToDeclarationProposal1(String label, ICompilationUnit targetCu, Object object, IMethodBinding methodDeclaration, int addSafevarargs) {
+		return new AddSafeVarargsProposal(label, targetCu, null, methodDeclaration, IProposalRelevance.ADD_SAFEVARARGS);
+	}
+	@Override
+	protected ICommandAccess createRemoveSafeVarargsProposal1(String label, ICompilationUnit compilationUnit, ASTRewrite rewrite, int removeSafevarargs) {
+		Image image= PlatformUI.getWorkbench().getSharedImages().getImage(ISharedImages.IMG_TOOL_DELETE);
+		return new ASTRewriteCorrectionProposal(label, compilationUnit, rewrite, IProposalRelevance.REMOVE_SAFEVARARGS, image);
+	}
 }
